+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * "src/res/img/bar/timerBar-0001.png"
 * "src/res/img/bar/json/timerBar-0001.json"
 * "src/res/img/character/teacher-0001.png"
+* "src/res/img/misc/correctnessIndicator.png"
+* "src/res/img/misc/json/correctnessIndicator.json"
 
 ### Updated