+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   * Updated version number
 * "src/main.cpp"
   * Updated version number
+  * Fixed warning
 * "src/font/GPixel.sfd"
   * Fixed space bar issue
 * "src/font/GPixel_v1.0.0.ttf"