+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 * "src/main.cpp"
   * Updated version number
   * Fixed warning
+  * Updated paths
 * "src/font/GPixel.sfd"
   * Fixed space bar issue
 * "src/font/GPixel_v1.0.0.ttf"