+++ CHANGELOG.md
@@ -26,6 +26,7 @@
   * Fixed warning
   * Updated paths
   * Added debug info
+  * Fixed path errors
 * "src/font/GPixel.sfd"
   * Fixed space bar issue
 * "src/font/GPixel_v1.0.0.ttf"