+++ CHANGELOG.md
@@ -56,6 +56,8 @@
   * Updated format
   * Added option for fullscreen mode
   * Changed default screen proportions
+  * Updated spritesheet logic to support new texture
+    * "src/res/img/misc/correctnessIndicator.png"
 
 ### Removed